+++ gha_utils/github/workflow_sync.py
@@ -32,8 +32,8 @@ from pathlib import Path
 
 import yaml
 
-from .init_project import export_content, get_data_content
-from .github import AnnotationLevel, emit_annotation
+from ..init_project import export_content, get_data_content
+from . import AnnotationLevel, emit_annotation
 
 if sys.version_info >= (3, 11):
     from enum import StrEnum